+++ windows/.env.example.ps1
@@ -2,7 +2,7 @@
 # You can overwrite this file in .env.ps1
 #
 
-$global:LNMP_PHP_IMAGE="khs1994/php:7.2.8-fpm-alpine"
+$global:LNMP_PHP_IMAGE="khs1994/php:7.2.9-fpm-alpine"
 
 $global:HYPERV_VIRTUAL_SWITCH='zy'
 
@@ -17,7 +17,7 @@ $COMMON_SOFT="nginx","php","mysql","wsl-redis"
 
 $global:NGINX_VERSION="1.15.2"
 # https://windows.php.net/download/
-$global:PHP_VERSION="7.2.8"
+$global:PHP_VERSION="7.2.9"
 $global:MYSQL_VERSION="8.0.12"
 $global:HTTPD_VERSION="2.4.34"
 $global:IDEA_VERSION="1.10.4088"