+++ wsl/lnmp-wsl-builder-php-rhel.sh
@@ -14,11 +14,11 @@ Build php in WSL RHEL by shell script
 
 Usage:
 
-$ lnmp-wsl-builder-php-rhel.sh 7.2.8
+$ lnmp-wsl-builder-php-rhel.sh 7.2.9
 
 $ lnmp-wsl-builder-php-rhel.sh yum
 
-$ lnmp-wsl-builder-php-rhel.sh 7.2.8 [--skipbuild] [tar] [rpm] [enable-ext]
+$ lnmp-wsl-builder-php-rhel.sh 7.2.9 [--skipbuild] [tar] [rpm] [enable-ext]
 
 "
 
@@ -691,7 +691,7 @@ command -v wget || sudo yum install -y wget
 
 mkdir -p /tmp/php-builder || echo
 
-if [ "$PHP_VERSION" = 'rpm' ];then PHP_VERSION=7.2.8 ; fi
+if [ "$PHP_VERSION" = 'rpm' ];then PHP_VERSION=7.2.9 ; fi
 
 _get_phpnum $PHP_VERSION