+++ .env.example.ps1
@@ -34,7 +34,7 @@ $DistributionName="debian"
 
 $COMMON_SOFT="nginx","php","mysql","wsl-redis"
 
-$global:NGINX_VERSION="1.15.2"
+$global:NGINX_VERSION="1.15.3"
 # https://windows.php.net/download/
 $global:PHP_VERSION="7.2.9"
 $global:MYSQL_VERSION="8.0.12"
@@ -43,7 +43,7 @@ $global:IDEA_VERSION="1.10.4088"
 $global:NODE_VEERSION="10.6.0"
 $global:GIT_VERSION="2.18.0"
 $global:PYTHON_VERSION="3.7.0"
-$global:GOLANG_VERSION="1.10.3"
+$global:GOLANG_VERSION="1.11"
 $global:HTTPD_MOD_FCGID_VERSION="2.3.9"
 $global:ZEAL_VERSION="0.5.0"