+++ lnmpkit/api/common.go
@@ -1,0 +1,5 @@
+package api // import "github.com/khs1994-docker/lnmp/lnmpkit/api"
+
+const (
+	DefaultVersion string = "18.06"
+)