+++ lnmpkit/cmd/lnmpkit/argv.go
@@ -1,0 +1,10 @@
+package main
+
+import "github.com/khs1994-docker/lnmp/lnmpkit/server/mysql"
+
+func Argv(argv []string){
+	switch argv[1] {
+	case "mysql":
+		mysql.Main(argv)
+	}
+}