+++ lnmpkit/server/mysql/helper.go
@@ -1,0 +1,7 @@
+package mysql
+
+import "fmt"
+
+func helper()  {
+	fmt.Println("server mysql help info")
+}