+++ windows/lnmp-k8s-dashboard.ps1
@@ -1,3 +1,0 @@
-cd $PSScriptRoot/../kubernetes
-
-kubectl apply -f kubernetes-dashboard.yaml