+++ CHANGELOG.md
@@ -7,7 +7,11 @@ Bug fixes:
 
 Changes:
 
+* Add Laravel Queue and Schedule Demo with systemd on Linux
+
 Updates:
+
+* PHP `7.2.9`
 
 #### [v18.09 (2018-08-21)](https://github.com/khs1994-docker/lnmp/releases/tag/v18.09)