+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changes:
 * Add Laravel Queue and Schedule Demo with systemd on Linux
 * Add `APP_ROOT` env, support set PHP Project folder
 * Add php nightly version,build from php/php-src git repo
+* Add data collection function, help us improve
 * Support `Minio`
 * Resupport PHP56 PHP57
 * Remove `gitlab` `gogs`, please see `khs1994-docker/ci`